actualizo logica de renderizado de sprite

+++ client/include/objetoJuego.h
@@ -19,7 +19,7 @@ public:
 
     void settear_estado(int posx, int posy);
 
-    void setDistanciaParcialAJugador(int& distancia);
+    void setDistanciaParcialAJugador(double& distancia);
 
     int getDistanciaParcialAJugador();
